move interfaces and edit imports for login pages
