+++ src/components/api/const/interfaces/loginInterfaces.tsx
@@ -1,0 +1,18 @@
+interface UserData {
+  accessToken: string;
+  name: string;
+  avatar?: { url: string; alt: string };
+  banner?: { url: string; alt: string };
+  bio?: string | null;
+  email: string;
+  venueManager: boolean;
+}
+
+export interface LoginFormInputs {
+  email: string;
+  password: string;
+}
+
+export interface LoginResponse {
+  data: UserData;
+}
